# docs/api/tcheck-corporate.openapi.yaml
openapi: 3.0.3
info:
  title: TallyCheck Corporate API
  description: |
    Corporate attendance, SafeChild pickup, BLE beacons and tenant settings.

    This is one of several TallyCheck service contracts. It does **not** cover the
    academic domain (courses, cohorts, enrolment) — that lives in `Tcheck-Backend`
    and is described by `tcheck-academic.openapi.yaml`. Visitor management lives in
    `vcheck.openapi.yaml`. The three are separate documents on purpose: they run on
    different hosts, use different auth, and have colliding paths (`/attendance`
    means corporate clock-in here and class attendance in the academic service).

    ## Multi-tenancy
    Every request resolves to exactly one tenant schema before any handler runs.
    See `shared/tenancy.yaml`. Callers holding a token send nothing extra — the
    token carries the tenant.

    ## Roles
    Roles are enumerated once, in `shared/primitives.yaml#/components/schemas/RoleKey`,
    and must match `libs/auth/src/roles.ts` and `backend/helpers/auth_helper.py`.
    Some tenants still carry the legacy names `hr` and `manager`; the backend
    guards accept both.

    ## Authentication
    All endpoints except `GET /health` and
    `GET /auth/organization-by-subdomain/{subdomain}` require a Bearer JWT.

    ## Response shape
    Successful responses are enveloped as `{ "message": ..., "data": ... }`, with
    `meta` added on paginated collections. See `shared/envelope.yaml`.
  version: 2.1.0
servers:
...
security:
  - Auth0: []

# ==========================================================
# REUSABLE STRUCTURES
# ==========================================================
components:
  securitySchemes:
    Auth0:
      type: openIdConnect
      # Must match the tenant configured in apps/intranet/src/app/app.config.ts.
      # This previously named a different tenant than the app actually used.
      openIdConnectUrl: https://dev-x506paw8afbw6jgu.us.auth0.com/.well-known/openid-configuration

  # ==========================================================
  # REUSABLE SCHEMAS
  # ==========================================================
  schemas:
...
    ErrorResponse:
      $ref: './shared/errors.yaml#/components/schemas/ErrorResponse'
...
    Beacon:
      type: object
      properties:
        id:
          $ref: './shared/primitives.yaml#/components/schemas/Uuid'
        name:
          type: string
          nullable: true
          example: Hekima Classroom Beacon
        mac_address:
          type: string
          description: Unique per tenant. The beacon identity.
          example: AC:23:45:67:89:01
        uuid:
          type: string
          nullable: true
          description: iBeacon proximity UUID.
        major:
          type: integer
          default: 1
        minor:
          type: integer
          default: 1
        zone_id:
          type: integer
          nullable: true
          description: Where the beacon is physically mounted.
        zone:
          # Null when the beacon has not been placed. OpenAPI 3.0 ignores a
          # `nullable` sibling next to `$ref`, so it is stated here instead.
          $ref: "#/components/schemas/ZoneRef"
        location:
          type: string
          nullable: true
          description: >
            Legacy free-text placement label, superseded by `zone`. Retained as a
            display fallback for rows the zone backfill could not match.
          example: Main Entrance
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          $ref: './shared/primitives.yaml#/components/schemas/Timestamp'
        updated_at:
          $ref: './shared/primitives.yaml#/components/schemas/Timestamp'
    BeaconRequest:
      type: object
      required: [mac_address]
      properties:
        name:
          type: string
        mac_address:
          type: string
          example: AC:23:45:67:89:01
        uuid:
          type: string
        major:
          type: integer
          default: 1
        minor:
          type: integer
          default: 1
        zone_id:
          type: integer
          nullable: true
        location:
          type: string
        description:
          type: string
        is_active:
          type: boolean
          default: true
    BeaconUpdateRequest:
      type: object
      description: Partial update — omitted fields are left unchanged.
      properties:
        name:
          type: string
        mac_address:
          type: string
        uuid:
          type: string
        major:
          type: integer
        minor:
          type: integer
        zone_id:
          type: integer
          nullable: true
        location:
          type: string
        description:
          type: string
        is_active:
          type: boolean
    BeaconAssignment:
      type: object
      properties:
        id:
          $ref: './shared/primitives.yaml#/components/schemas/Uuid'
        beacon_id:
          $ref: './shared/primitives.yaml#/components/schemas/Uuid'
        beacon_name:
          type: string
          nullable: true
        beacon_mac:
          type: string
          nullable: true
        department_id:
          type: integer
          nullable: true
        department_name:
          type: string
          nullable: true
          description: Prettified from the stored slug.
          example: Software Development
        created_at:
          $ref: './shared/primitives.yaml#/components/schemas/Timestamp'

    # ==========================================================
    # SAFECHILD
    # ==========================================================
...
    ZoneType:
      type: string
      description: >
        Physical kind of place. Like `UnitType` this is a fixed structural list;
        customer-specific wording is a display concern.
      enum:
        - room
        - classroom
        - hall
        - entrance
        - gate
        - loading_bay
        - storage
        - ward
        - yard
        - general
      default: general
    ZoneRef:
      type: object
      description: Compact zone, as nested inside a beacon.
      properties:
        id:
          type: integer
        name:
          type: string
          example: Loading Bay 3
        zone_type:
          $ref: "#/components/schemas/ZoneType"
...
  responses:
    Unauthorized: # Can be referenced as '#/components/responses/unauthorized'
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    UnprocessableEntity:
      description: Validation error e.g. missing required field
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ==========================================================
  # REUSABLE PARAMETERS
  # ==========================================================
...
paths:
...
  /beacons:
    get:
      tags: [Beacons]
      summary: List BLE beacons
      responses:
        "200":
          description: Beacons in this tenant, ordered by name.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Beacon"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Beacons]
      summary: Register a beacon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeaconRequest"
      responses:
        "201":
          description: Beacon registered.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Beacon"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /beacons/{beacon_id}:
    parameters:
      - name: beacon_id
        in: path
        required: true
        schema:
          $ref: './shared/primitives.yaml#/components/schemas/Uuid'
    put:
      tags: [Beacons]
      summary: Update a beacon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeaconUpdateRequest"
      responses:
        "200":
          description: Beacon updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Beacon"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    delete:
      tags: [Beacons]
      summary: Delete a beacon
      description: Cascades to its assignments.
      responses:
        "200":
          description: Beacon deleted.
          content:
            application/json:
              schema:
                $ref: './shared/envelope.yaml#/components/schemas/DeletedEnvelope'
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /beacons/assignments:
    get:
      tags: [Beacons]
      summary: List beacon-to-department assignments
      responses:
        "200":
          description: All assignments in this tenant.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/BeaconAssignment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Beacons]
      summary: Assign a beacon to a department
      description: A beacon may serve several departments; the pair must be unique.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [beacon_id, department_id]
              properties:
                beacon_id:
                  $ref: './shared/primitives.yaml#/components/schemas/Uuid'
                department_id:
                  type: integer
      responses:
        "201":
          description: Assignment created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/BeaconAssignment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /beacons/assignments/{assignment_id}:
    delete:
      tags: [Beacons]
      summary: Remove a beacon assignment
      parameters:
        - name: assignment_id
          in: path
          required: true
          schema:
            $ref: './shared/primitives.yaml#/components/schemas/Uuid'
      responses:
        "200":
          description: Assignment removed.
          content:
            application/json:
              schema:
                $ref: './shared/envelope.yaml#/components/schemas/DeletedEnvelope'
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==========================================================
  # SAFECHILD
  # ==========================================================
...