# libs/api-client/src/generated/tcheck/model/beaconAssignment.ts
/**
 * Generated by orval from TallyCheck Corporate API v2.1.0.
 * DO NOT EDIT — run `nx run api-client:generate` instead.
 */
import type { Timestamp } from "./timestamp";
import type { Uuid } from "./uuid";

export interface BeaconAssignment {
  id?: Uuid;
  beacon_id?: Uuid;
  /** @nullable */
  beacon_name?: string | null;
  /** @nullable */
  beacon_mac?: string | null;
  /** @nullable */
  department_id?: number | null;
  /**
   * Prettified from the stored slug.
   * @nullable
   */
  department_name?: string | null;
  created_at?: Timestamp;
}