# libs/api-client/src/generated/tcheck/model/zoneRef.ts
/**
 * Generated by orval from TallyCheck Corporate API v2.1.0.
 * DO NOT EDIT — run `nx run api-client:generate` instead.
 */
import type { ZoneType } from "./zoneType";

/**
 * Compact zone, as nested inside a beacon.
 */
export interface ZoneRef {
  id?: number;
  name?: string;
  zone_type?: ZoneType;
}